changes CRUD functionality for admin escort according to media table

+++ app/Http/Controllers/AdminEscortsController.php
@@ -78,7 +78,8 @@ class AdminEscortsController extends Controller
         if ($escort_id) {
             // Pictures upload in media table
             if (is_array($request->file('pictures'))) {
-                // Handle multiple files
+                // Handle multiple files           
+
                 foreach ($request->file('pictures') as $image) {
                     $originalImageName = $image->getClientOriginalName();
                     $imageName = time() . '_' . $originalImageName;
@@ -89,18 +90,6 @@ class AdminEscortsController extends Controller
                         'escort_id' => $escort_id,
                     ]);
                 }
-            } else {
-                // Handle single file
-                $image = $request->file('name');
-                $originalImageName = $image->getClientOriginalName();
-                $imgName = time() . '_' . $originalImageName;
-                $image->move(public_path('images/escorts_img'), $imgName);
-
-                Media::create([
-                    'name' => $imgName,
-                    'type' => $request->image_type,
-                    'escort_id' => $escort_id,
-                ]);
             }
 
 
@@ -117,17 +106,6 @@ class AdminEscortsController extends Controller
                         'escort_id' => $escort_id,
                     ]);
                 }
-            } else {
-                // Handle single file
-                $vdo = $request->file('name');
-                $originalVdoName = $vdo->getClientOriginalName();
-                $vdoName = time() . '_' . $originalVdoName;
-                $vdo->move(public_path('videos'), $vdoName);
-                Media::create([
-                    'name' => $vdoName,
-                    'type' => $request->video_type,
-                    'escort_id' => $escort_id,
-                ]);
             }
         }
 
@@ -137,21 +115,28 @@ class AdminEscortsController extends Controller
     public function edit_escorts_form($id)
     {
         $escort = Escort::find($id);
+        // Check if escort was found
+        if (!$escort) {
+            return redirect()->back()->with('error', 'Escort not found.');
+        }
+        $pictures = Media::where('escort_id', $id)->where('type', 'image')->get();
+        $videos = Media::where('escort_id', $id)->where('type', 'video')->get();
 
-        $pictures = json_decode($escort->pictures);
-        $video = json_decode($escort->video);
         $language_spoken = json_decode($escort->language_spoken, true);
         $services = json_decode($escort->services, true);
         $availability = json_decode($escort->availability, true);
         $currencies_accepted = json_decode($escort->currencies_accepted, true);
         $payment_method = json_decode($escort->payment_method, true);
 
-        return view('escorts.escorts-edit', compact('escort', 'pictures', 'video', 'services', 'language_spoken', 'availability', 'currencies_accepted', 'payment_method'));
+        return view('escorts.escorts-edit', compact('escort', 'pictures', 'videos', 'services', 'language_spoken', 'availability', 'currencies_accepted', 'payment_method'));
     }
     //todo: update Escorts
     public function edit_escorts(Request $request, $id)
     {
         $escort = Escort::findOrFail($id);
+        if (!$escort) {
+            return redirect()->back()->with('error', 'Escort not found.');
+        }
 
         $validatedData = $request->validate([
             'nickname' => 'required|unique:escorts,nickname,' . $escort->id,
@@ -165,8 +150,8 @@ class AdminEscortsController extends Controller
             'origin' => 'required',
             'type' => 'required',
             'text_description' => 'required|min:30',
-            'video' => 'nullable|array',
-            'video.*' => 'file|mimes:mp4,mov,mkv,flv,3gp,avi,mwv,ogg,qt|max:20000',
+            'videos' => 'nullable|array',
+            'videos.*' => 'file|mimes:mp4,mov,mkv,flv,3gp,avi,mwv,ogg,qt|max:20000',
             'hair_color' => 'nullable',
             'hair_length' => 'nullable',
             'breast_size' => 'nullable',
@@ -188,37 +173,47 @@ class AdminEscortsController extends Controller
             'rates_in_chf' => 'nullable|numeric',
             'currencies_accepted' => 'nullable|array',
             'payment_method' => 'nullable|array',
+            'image_type' => 'nullable|string',
+            'video_type' => 'nullable|string',
         ]);
 
-        // Handle Image file upload
-        $pictures = json_decode($escort->pictures, true) ?? [];
-        if ($request->hasFile('pictures')) {
-            foreach ($request->file('pictures') as $image) {
-                $originalImageName = $image->getClientOriginalName();
-                $imageName = time() . '_' . $originalImageName;
-                $image->move(public_path('images/escorts_img'), $imageName);
-                $pictures[] = $imageName;
-            }
-        }
 
-        // Handle video file upload
-        $videos = json_decode($escort->video, true) ?? [];
-        if ($request->hasFile('video')) {
-            foreach ($request->file('video') as $vdo) {
-                $originalVdoName = $vdo->getClientOriginalName();
-                $vdoName = time() . '_' . $originalVdoName;
-                $vdo->move(public_path('videos'), $vdoName);
-                $videos[] = $vdoName;
-            }
-        }
-
-        $validatedData['pictures'] = json_encode($pictures);
         $validatedData['services'] = json_encode($validatedData['services']);
-        $validatedData['video'] = json_encode($videos);
         $validatedData['language_spoken'] = isset($validatedData['language_spoken']) ? json_encode($validatedData['language_spoken']) : null;
         $validatedData['availability'] = isset($validatedData['availability']) ? json_encode($validatedData['availability']) : null;
         $validatedData['currencies_accepted'] = isset($validatedData['currencies_accepted']) ? json_encode($validatedData['currencies_accepted']) : null;
         $validatedData['payment_method'] = isset($validatedData['payment_method']) ? json_encode($validatedData['payment_method']) : null;
+
+        // Pictures upload in media table
+        if (is_array($request->file('pictures'))) {
+            // Handle multiple files
+            foreach ($request->file('pictures') as $image) {
+                $originalImageName = $image->getClientOriginalName();
+                $imageName = time() . '_' . $originalImageName;
+                $image->move(public_path('images/escorts_img'), $imageName);
+                Media::create([
+                    'name' => $imageName,
+                    'type' => $request->image_type,
+                    'escort_id' => $id,
+                ]);
+            }
+        }
+
+
+        // Videos upload in media table
+        if (is_array($request->file('videos'))) {
+            // Handle multiple files
+            foreach ($request->file('videos') as $vdo) {
+                $originalVdoName = $vdo->getClientOriginalName();
+                $vdoName = time() . '_' . $originalVdoName;
+                $vdo->move(public_path('videos'), $vdoName);
+                Media::create([
+                    'name' => $vdoName,
+                    'type' => $request->video_type,
+                    'escort_id' => $id,
+                ]);
+            }
+        }
 
         $escort->update($validatedData);
 
@@ -230,6 +225,9 @@ class AdminEscortsController extends Controller
     public function escorts_by_id($id)
     {
         $escorts = Escort::find($id);
+        if (!$escorts) {
+            return redirect()->back()->with('error', 'Escort not found.');
+        }
 
         $pictures = Media::where('escort_id', $id)->where('type', 'image')->get();
         $videos = Media::where('escort_id', $id)->where('type', 'video')->get();
@@ -245,31 +243,41 @@ class AdminEscortsController extends Controller
     //todo: Delete Escorts
     public function deleteEscorts($id)
     {
-        $escorts = Escort::find($id);
-        $pictures = json_decode($escorts->pictures);
-        $video = json_decode($escorts->video);
+        $escort = Escort::find($id);
 
+        if (!$escort) {
+            return redirect()->route('admin.escorts')->with('error', 'Escort not found.');
+        }
+
+        // Find and delete all media records related to the escort        
+        $pictures = Media::where('escort_id', $id)->where('type', 'image')->get();
+        $videos = Media::where('escort_id', $id)->where('type', 'video')->get();
         //delete escorts pics
         if ($pictures) {
             foreach ($pictures as $picture) {
-                $imagePath = public_path('images/escorts_img') . '/' . $picture;
+                $imagePath = public_path('images/escorts_img') . '/' . $picture->name;
                 if (file_exists($imagePath)) {
                     unlink($imagePath);
                 }
+
+                $picture->delete();
             }
         }
 
         //Delete Escorts videos
-        if ($video) {
-            foreach ($video as $vdo) {
-                $vdoPath = public_path('videos') . '/' . $vdo;
+        if ($videos) {
+            foreach ($videos as $video) {
+                $vdoPath = public_path('videos') . '/' . $video->name;
                 if (file_exists($vdoPath)) {
                     unlink($vdoPath);
                 }
+
+                $video->delete();
             }
         }
 
-        $escorts->delete();
+
+        $escort->delete();
         return redirect()->route('admin.escorts')->with('success', 'Escorts deleted successfully');
     }
 }
